CREATE db/models/product_footprint_model.py - add model with basic tests
This model is so simple that it's almost a joke, I reduced the complexity of this
module a while ago just to get shit working, and have continued on that way since
then. I will be undertaking a serious look at the data model during beta and
seriously ripping it apart and isolating things in a much better way.