CREATE EmissionFactorDS entity - add datamodel, along with type alias for NonEmptyString
The specification and the data model deviated in terms of what was expected for
the secondaryEmissionFactorSources field. A list of str was being provided, which
was incorrect. This represents the introduction of the new type, along with a base
type that is a type alias of string to implement NonEmptyString, in order to
more closely align with the specification. The next commit will be the introduction
and rewriting of module code and tests to reflect this change.